admin_panel/requirements.txt
Django==4.1.6
django-split-settings==1.0.0
python-dotenv==0.21.1
gunicorn==20.1.0